split client logic in slicer into seperate module
